removed defaultpersistentstore from lib, added to tests
