+++ test/base/controller_test.dart
@@ -7,6 +7,7 @@ import 'package:aqueduct/aqueduct.dart';
 import 'dart:convert';
 import 'dart:async';
 import 'dart:mirrors';
+import '../helpers.dart';
 
 void main() {
   HttpServer server;
